Extract baseline handling code
